+++ gradle.properties
@@ -1,0 +1,7 @@
+VERSION_NAME=1.0
+VERSION_CODE=1
+
+ANDROID_MIN_SDK=8
+ANDROID_TARGET_SDK=19
+ANDROID_COMPILE_SDK=19
+ANDROID_BUILD_TOOLS_VERSION=19.0.3
