+++ library/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest package="com.almeros.android.multitouch">
+
+    <application />
+
+</manifest>
