+++ sample/src/main/AndroidManifest.xml
@@ -1,23 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.almeros.android.multitouch"
-    android:versionCode="1"
-    android:versionName="1.0" >
-
-    <uses-sdk
-        android:minSdkVersion="8"
-        android:targetSdkVersion="18" />
+    package="com.almeros.android.multitouch.sample">
 
     <uses-configuration android:reqTouchScreen="finger" />
 
     <uses-feature
         android:name="android.hardware.touchscreen"
         android:required="true" />
+
     <uses-feature
         android:name="android.hardware.touchscreen.multitouch"
         android:required="true" />
     
-
     <application
         android:icon="@drawable/earth"
         android:label="@string/app_name"
