+++ sample/src/main/res/drawable/basketball.png
