+++ sample/src/main/res/drawable/earth.png
