+++ sample/src/main/res/drawable/stars.png
