+++ sample/src/main/res/layout/main.xml
