+++ sample/src/main/res/values/strings.xml
